requirements.txt
discord.py
pynacl
yt_dlp
python-dotenv
homeassistant-api
pytest